feat: refactor again
