+++ wrappers/src/fdw/auth0_fdw/auth0_client/row.rs
@@ -1,5 +1,29 @@
+use pgrx::JsonB;
+use serde::Deserialize;
+use serde_json::Value;
+use std::collections::HashMap;
+use supabase_wrappers::prelude::Cell;
+use supabase_wrappers::prelude::Column;
+use supabase_wrappers::prelude::Row;
+
+#[derive(Deserialize, Debug)]
+pub struct Auth0Response {
+    pub records: Vec<Auth0User>,
+}
+
+#[derive(Debug)]
+pub struct Auth0Fields(HashMap<String, Value>);
+
+#[derive(Deserialize, Debug)]
+pub struct Auth0User {
+    pub created_at: String,
+    pub email: String,
+    pub email_verified: bool,
+    pub identities: Option<serde_json::Value>,
+}
+
 impl Auth0User {
-    pub(crate) fn to_row(&self, columns: &[Column]) -> Auth0FdwResult<Row> {
+    pub(crate) fn to_row(mut self, columns: &[Column]) -> Row {
         let mut row = Row::new();
         for tgt_col in columns {
             if tgt_col.name == "created_at" {
@@ -12,12 +36,15 @@ impl Auth0User {
             } else if tgt_col.name == "email_verified" {
                 row.push("email_verified", Some(Cell::Bool(self.email_verified)))
             } else if tgt_col.name == "identities" {
-                let attrs = serde_json::from_str(&self.identities.to_string())?;
+                let attrs = self
+                    .identities
+                    .take()
+                    .expect("Column `identities` missing in response");
 
                 row.push("identities", Some(Cell::Json(JsonB(attrs))))
             }
         }
 
-        Ok(row)
+        row
     }
 }
